requirements.txt
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
scipy==1.4.1
python-igraph==0.8.2
